Fix golangci-lint security and style issues
- G602: Add bounds check in racing provider pickBestCandidate
- G115: Use uint64 type for numGoroutines in loadbalance test
- G306: Change file permissions from 0644 to 0600 in fileutils test
- S1024: Use time.Until() instead of t.Sub(time.Now())
- SA1029: Add contextKey type for context values
- SA4006: Move registry init after t.Skip() in models test
- SA5011: Use t.Fatal() with early return in config test

+++ pkg/providers/virtual/loadbalance/loadbalance_test.go
@@ -409,10 +409,10 @@ func TestRoundRobinStrategy_ConcurrentAccess(t *testing.T) {
 	opts := types.GenerateOptions{Prompt: "test"}
 
 	var wg sync.WaitGroup
-	numGoroutines := 100
+	numGoroutines := uint64(100)
 	errChan := make(chan error, numGoroutines)
 
-	for i := 0; i < numGoroutines; i++ {
+	for i := uint64(0); i < numGoroutines; i++ {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
@@ -433,7 +433,7 @@ func TestRoundRobinStrategy_ConcurrentAccess(t *testing.T) {
 	}
 
 	// Counter should be exactly numGoroutines
-	if lb.counter != uint64(numGoroutines) {
+	if lb.counter != numGoroutines {
 		t.Errorf("expected counter to be %d, got %d", numGoroutines, lb.counter)
 	}
 }
